+++ python/python/tests/test_transaction.py
@@ -220,6 +220,8 @@ async def transaction_test(
         },
     )
     args.append(2)
+    transaction.geodist(key9, "Palermo", "Catania")
+    args.append(166274.1516)
     transaction.geohash(key9, ["Palermo", "Catania", "Place"])
     args.append(["sqc8b49rny0", "sqdtr74hyu0", None])
     return args